+++ test-suite/Main.hs
@@ -77,6 +77,7 @@ main = do
       -- , golden "specialization-2-py - bytes/bytearray" "specialization-2-py"
       -- , golden "specialization-1-r" "specialization-1-r"
 
+      , golden "int-literal-promoted-to-real" "int-literal-promoted-to-real"
       , golden "backtick-operator-py" "backtick-operator-py"
       , golden "crosslang-io-thunk-wrap" "crosslang-io-thunk-wrap"
       , golden "effect-boundary-py" "effect-boundary-py"